if (currentItem == null)             {                 //log                 //make it more MVVM friendly                 MessageBox.Show("FEHLER: Bitte wählen Sie zuerst eine Tour aus!");                 return;             }

+++ TourPlanner_Lercher_Polley/ViewModels/TourPlannerViewModel.cs
@@ -147,6 +147,13 @@ namespace TourPlanner_Lercher_Polley.ViewModels
         }
         private void deleteTour(object commandParameter)
         {
+            if (currentItem == null)
+            {
+                //log
+                //make it more MVVM friendly
+                MessageBox.Show("FEHLER: Bitte wählen Sie zuerst eine Tour aus!");
+                return;
+            }
             tourManager.deleteTour(currentItem);
             Items.Clear();
             LoadList();
@@ -154,6 +161,13 @@ namespace TourPlanner_Lercher_Polley.ViewModels
 
         private void editTour(object commandParameter)
         {
+            if (currentItem == null)
+            {
+                //log
+                //make it more MVVM friendly
+                MessageBox.Show("FEHLER: Bitte wählen Sie zuerst eine Tour aus!");
+                return;
+            }
             AddTourWindow editTourWindow = new AddTourWindow((int)currentItem.ID,currentItem.Name, currentItem.To, currentItem.From,
                 currentItem.Description, currentItem.TransportType);            
             editTourWindow.ShowDialog();
@@ -180,6 +194,13 @@ namespace TourPlanner_Lercher_Polley.ViewModels
 
         public void deleteTourLog(object commandParameter)
         {
+            if (currentItem == null)
+            {
+                //log
+                //make it more MVVM friendly
+                MessageBox.Show("FEHLER: Bitte wählen Sie zuerst eine Tour aus!");
+                return;
+            }
             logManager.deleteLog(CurrentLog);
             Items.Clear();
             LoadList();
@@ -187,6 +208,7 @@ namespace TourPlanner_Lercher_Polley.ViewModels
 
         public void editTourLog(object commandParameter)
         {
+
             if (currentItem == null)
             {
                 //log
